moved percentage inputs to Percentage component, updated handling functions, updated imports
