+++ src/components/index.js
@@ -1,6 +1,7 @@
 export * from './Buttons';
 export * from './Footer';
 export * from './Navbar';
+export * from './Percentage';
 export * from './SessionData';
 export * from './SessionType';
 export * from './StudentObs';
